Admin clearlimit, and error reporting.

+++ cogs/admin.py
@@ -2,9 +2,11 @@ import discord
 import asyncio
 
 from library.ConsoleLib import Time
+from library.ConsoleLib import Color
 from library.ConsoleLib import DurationConverter
 from library.ConsoleLib import Roundtrip
 
+from settings.Settings import Admin
 from discord.ext import commands
 
 class Administration(commands.Cog):
@@ -16,14 +18,26 @@ class Administration(commands.Cog):
     # CLEAR METHOD
     @commands.command(help = "Clears an amount of messages.")
     @commands.has_permissions(manage_messages = True)
+    @commands.cooldown(rate = 1, per = 5)
     async def clear(self, ctx, amount = 5):
-        message = discord.Message
-        await message.delete()
-        await ctx.channel.purge(limit = amount)
-        print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - CLEAR: A channel was cleared using this command!')
+        if (amount >= Admin.clearlimit):
+            embed = discord.Embed (
+                description = f'Woah! too many requests! Try something less than {Admin.clearlimit}.',
+                color = Color.tachi
+            )
+            embed.set_footer (
+                text = f'Tachibana: {Time.timeFormatUniversial()}'
+            )            
+            message: discord.Message = await ctx.reply(embed = embed)
+            await asyncio.sleep(3)
+            await message.delete()
+            print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - CLEAR: Channel #{ctx.channel} was unsuccessfully cleared! Larger amount than expected! [Amount: {amount}] [Limit: {Admin.clearlimit}]')
+        else:
+            await ctx.channel.purge(limit = amount)
+            print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - CLEAR: Channel #{ctx.channel} was cleared {amount} times using this command!')
     
     @clear.error
-    async def clear_error(ctx, error):
+    async def clear_error(self, ctx, error):
         if isinstance(error, commands.MissingRequiredArgument):
             await ctx.reply('Please specify an amount of messages to delete.')
 
@@ -36,7 +50,7 @@ class Administration(commands.Cog):
         else:
             await member.kick(reason = reason)
             await ctx.reply(f'Kicked **{member}!**')
-            print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - KICK: A user was kicked! {member}')
+            print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - KICK: A user was kicked in #{ctx.channel}! [{member}]')
         
     @kick.error
     async def kick_error(ctx, error):
@@ -52,7 +66,7 @@ class Administration(commands.Cog):
         else:    
             await ctx.guild.ban(member, reason = reason)
             await ctx.reply(f'Banned **{member}!**')
-            print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - BAN: A user was banned! {member}')
+            print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - BAN: A user was banned in #{ctx.channel}! [{member}]')
     
     # TEMP BAN METHOD
     @commands.command(help = "Temporarily bans a person from the server. - (UNSTABLE)")
@@ -68,7 +82,7 @@ class Administration(commands.Cog):
             await ctx.reply(f'**{member}!** has been banned for **{amount}{unit}.**')
             await asyncio.sleep(amount * multiplier[unit])
             await ctx.guild.unban(member)
-            print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - TEMPBAN: A user was temporarily banned! {member}')
+            print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - TEMPBAN: A user was temporarily banned in #{ctx.channel}! [{member}]')
 
     # UNBAN METHOD
     @commands.command(help = "Unbans a person from the server.")
@@ -87,7 +101,7 @@ class Administration(commands.Cog):
                     await ctx.guild.unban(user)
                     await ctx.reply(f'Unbanned **{user.mention}!**')
                     
-                    print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - UNBAN: A user was unbanned! {member}')
+                    print(f'[{Time.timeFormat()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - UNBAN: A user was unbanned in #{ctx.channel}! [{member}]')
                     return  
 
 def setup(client):
